Add ublox-wrap
ublox-wrap is a simple tcp wrapper of ublox connected via uart.
Uart tty output from ublox will be copied to tcp and udp ports.  TCP
inputs are directed to uart and UDP inputs are ignored.  UDP port is
intended to be used as the source for something like GPS daemon.